AI-ChatBot/requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_huggingface
python-dotenv

# .venv\Scripts\activate

#python -m uvicorn main:app --host 0.0.0.0 --port 8000 --reload
#10.0.2.2 - IP address of the host machine


# http://10.0.2.2:8000/query

